requirements.txt
streamlit
google-generativeai
langchain
pypdf2
python-dotenv
chromadb
pymupdf